BlobModel now uses doubles

+++ src/org/ilite/vision/camera/tools/colorblob/BlobModel.java
@@ -1,18 +1,23 @@
 package org.ilite.vision.camera.tools.colorblob;
 
+import java.io.File;
+
+import javax.xml.bind.JAXBException;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
+
+import org.ilite.vision.data.XMLManager;
 
 @XmlRootElement
 public class BlobModel {
     private String name;
-    private float[] hsv;
+    private double[] hsv;
     private static final int HUE = 0;
     private static final int SATURATION = 1;
     private static final int VALUE = 2;
     
     public BlobModel() {
-        hsv = new float[3];
+        hsv = new double[3];
     }
     
     @XmlElement
@@ -21,29 +26,29 @@ public class BlobModel {
     }
     
     @XmlElement
-    public void setAverageHue(float hue) {
+    public void setAverageHue(double hue) {
         hsv[HUE] = hue;
     }
     
     @XmlElement
-    public void setAverageValue(float value) {
+    public void setAverageValue(double value) {
         hsv[VALUE] = value;
     }
     
     @XmlElement
-    public void setAverageSaturation(float saturation) {
+    public void setAverageSaturation(double saturation) {
         hsv[SATURATION] = saturation;
     }
     
-    public float getAverageHue() {
+    public double getAverageHue() {
         return hsv[HUE];
     }
 
-    public float getAverageValue() {
+    public double getAverageValue() {
         return hsv[VALUE];
     }
 
-    public float getAverageSaturation() {
+    public double getAverageSaturation() {
         return hsv[SATURATION];
     }
